+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 - Improve developement instructions (@marteinn)
 - Make sure title and descriptions are truncated if too long (@marteinn)
+- Strip html tags from preview widget (@marteinn)
 
 ### Fixed
 - Fix broken image preview #12 (@marteinn)